Move OctetsExtensionsTests
