+++ 2-test/networking.model.tests/ExtensionsTests/OctetsExtensionsTests/Octets_Method_SetIPv6_Test.cs
@@ -3,7 +3,7 @@ using FluentAssertions;
 using Networking.Model.Internet;
 using Xunit;
 
-namespace Networking.Model.Tests.ExtensionsTests
+namespace Networking.Model.Tests.ExtensionsTests.OctetsExtensionsTests
 {
     public class Octets_Method_SetIPv6_Test
     {
